// packages/replay-component/src/types/simulation.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type SimulationEvent =
  | HeaderEvent
  | FooterEvent
  | BallCollisionEvent
  | PaddleFrameEvent
  | TurnStartEvent
  | TurnEndEvent
/**
 * JSON Schema for Pong game simulation events
 */
export type SimulationEvents = SimulationEvent[]

export interface HeaderEvent {
  event_type: 'HeaderEvent'
  version: '0.0.1'
  game: {
    width: number
    height: number
    paddle_l: PaddleProps
    paddle_r: PaddleProps
    [k: string]: unknown
  }
  [k: string]: unknown
}
export interface PaddleProps {
  width: number
  [k: string]: unknown
}
export interface FooterEvent {
  event_type: 'FooterEvent'
  [k: string]: unknown
}
export interface BallCollisionEvent {
  event_type: 'BallCollisionEvent'
  frame_index: number
  position: Vec2D
  collision_type: 'top' | 'bottom' | 'left' | 'right' | 'paddle_l' | 'paddle_r'
  [k: string]: unknown
}
export interface Vec2D {
  x: number
  y: number
  [k: string]: unknown
}
export interface PaddleFrameEvent {
  event_type: 'PaddleFrameEvent'
  frame_index: number
  paddle_l: PaddleFrameState
  paddle_r: PaddleFrameState
  [k: string]: unknown
}
export interface PaddleFrameState {
  force?: Vec2D
  position: Vec2D
  velocity: Vec2D
  endurance: number
  [k: string]: unknown
}
export interface TurnStartEvent {
  event_type: 'TurnStartEvent'
  turn_index: number
  [k: string]: unknown
}
export interface TurnEndEvent {
  event_type: 'TurnEndEvent'
  result: 'left' | 'right' | 'draw'
  turn_index: number
  [k: string]: unknown
}